requirements.txt
aiohttp==3.7.4.post0
alembic==1.8.1
aniso8601==9.0.1
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==3.0.1
attrs==22.1.0
awsebcli==3.20.3
bcrypt==4.0.0
beautifulsoup4==4.11.1
blessed==1.19.1
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cmdstanpy==1.0.8
colorama==0.4.3
convertdate==2.4.0
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.1
django-scheduler==0.10.0
dnspython==2.2.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.2.1
ephem==4.1.3
fake-useragent==0.1.11
Faker==13.15.1
feedparser==6.0.8
Flask==2.2.2
Flask-Compress==1.12
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-Minify==0.39
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.37.1
future==0.16.0
greenlet==1.1.3
gunicorn==20.1.0
hijri-converter==2.2.4
holidays==0.16
htmlmin==0.1.12
icalendar==4.1.0
idna==3.3
importlib-metadata==4.11.3
iohandler==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
jsmin==3.0.1
jsonschema==3.2.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lesscpy==0.15.0
LunarCalendar==0.0.9
lxml==4.8.0
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
multitasking==0.0.10
nltk==3.7
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pandas-datareader==0.10.0
paramiko==2.11.0
parse==1.19.0
pathlib==1.0.1
pathspec==0.9.0
Pillow==9.2.0
plotly==5.10.0
ply==3.11
praw==7.6.0
prawcore==2.3.0
prophet==1.1.1
psycopg==3.0.14
psycopg2-binary==2.9.3
pycoingecko==2.2.0
pycparser==2.21
pyee==8.2.2
PyMeeus==0.5.11
PyNaCl==1.5.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-edgar==3.1.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
rake-nltk==1.0.6
rcssmin==1.1.1
regex==2022.7.25
renderer==1.0.3
requests==2.26.0
requests-html==0.10.0
s3transfer==0.6.0
schedule==1.1.0
scikit-learn==1.1.3
scipy==1.9.1
seaborn==0.11.2
sec-edgar-downloader==4.3.0
secedgar==0.4.0
semantic-version==2.8.5
setuptools-git==1.2
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
structlog==22.1.0
tenacity==8.0.1
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==3.1.0
tqdm==4.64.0
typing==3.7.4.3
typing-extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.9
w3lib==1.22.0
wcwidth==0.1.9
websocket-client==0.59.0
websockets==10.3
Werkzeug==2.2.2
WTForms==3.0.1
xxhash==3.0.0
yahoo-fin==0.8.9.1
yarl==1.7.2
yfinance==0.1.70
zipp==3.8.0
